move metadata providers its data structures to libcst.metadata
