+++ libcst/_metadata_dependent.py
