+++ libcst/_visitors.py
@@ -6,10 +6,10 @@
 # pyre-strict
 from typing import TYPE_CHECKING, Union
 
+from libcst._metadata_dependent import MetadataDependent
 from libcst._removal_sentinel import RemovalSentinel
 from libcst._typed_visitor import CSTTypedTransformerFunctions, CSTTypedVisitorFunctions
 from libcst._types import CSTNodeT
-from libcst.metadata.dependent import MetadataDependent
 
 
 if TYPE_CHECKING:
