+++ libcst/metadata/__init__.py
@@ -4,3 +4,42 @@
 # LICENSE file in the root directory of this source tree.
 
 # pyre-strict
+
+from libcst.metadata.expression_context_provider import (
+    ExpressionContext,
+    ExpressionContextProvider,
+)
+from libcst.metadata.position_provider import (
+    BasicPositionProvider,
+    SyntacticPositionProvider,
+)
+from libcst.metadata.scope_provider import (
+    Access,
+    Assignment,
+    BaseAssignment,
+    BuiltinAssignment,
+    ClassScope,
+    ComprehensionScope,
+    FunctionScope,
+    GlobalScope,
+    Scope,
+    ScopeProvider,
+)
+
+
+__all__ = [
+    "BasicPositionProvider",
+    "SyntacticPositionProvider",
+    "ExpressionContext",
+    "ExpressionContextProvider",
+    "BaseAssignment",
+    "Assignment",
+    "BuiltinAssignment",
+    "Access",
+    "Scope",
+    "GlobalScope",
+    "FunctionScope",
+    "ClassScope",
+    "ComprehensionScope",
+    "ScopeProvider",
+]
